[MIG] Migrate to v8
