+++ l10n_br_account_withholding/__openerp__.py
@@ -1,23 +1,7 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    Copyright (C) 2014 KMEE (http://www.kmee.com.br)
-#    @author Luis Felipe Mileo <mileo@kmee.com.br>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# -*- coding: utf-8 -*-
+# © 2016 KMEE(http://www.kmee.com.br)
+#   @author Luis Felipe Mileo <mileo@kmee.com.br>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 {
     'name': 'Brazilian Localization Account withholdings',
@@ -25,16 +9,17 @@
     Brazilian Localization Account withholdings""",
     'category': 'Localisation',
     'license': 'AGPL-3',
-    'author': 'KMEE',
+    'author': 'KMEE, Odoo Community Association (OCA)',
     'website': 'http://www.kmee.com.br',
-    'version': '7.0',
+    'version': '8.0.1.0.1',
     'depends': [
-        'l10n_br_account_product',
+        'l10n_br_account_product_service',
+        'l10n_br_sale',
     ],
     'data': [
-        'account_invoice_view.xml',
-        'l10n_br_account_view.xml',
-        'res_company_view.xml',
+        'views/account_invoice_view.xml',
+        'views/l10n_br_account_view.xml',
+        'views/res_company_view.xml',
     ],
-    'installable': True
+    'installable': True,
 }
