+++ l10n_br_account_withholding/models/__init__.py
@@ -3,4 +3,7 @@
 #   @author Luis Felipe Mileo <mileo@kmee.com.br>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from . import models
+from . import account_invoice
+from . import l10n_br_account
+from . import res_company
+from . import sale_order
