+++ l10n_br_account_withholding/models/account_invoice.py
@@ -1,23 +1,7 @@
-# -*- encoding: utf-8 -*-
-# #############################################################################
-#
-# Copyright (C) 2014 KMEE (http://www.kmee.com.br)
-# @author Luis Felipe Mileo <mileo@kmee.com.br>
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# -*- coding: utf-8 -*-
+# © 2016 KMEE(http://www.kmee.com.br)
+#   @author Luis Felipe Mileo <mileo@kmee.com.br>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 
 from openerp import api, fields, models
@@ -34,7 +18,7 @@ class AccountInvoice(models.Model):
     def _amount_all_service(self):
         for inv in self:
             inv.amount_services = sum(
-                line.price_total for line in inv.invoice_line)
+                line.price_subtotal for line in inv.invoice_line)
             inv.issqn_base = sum(line.issqn_base for line in inv.invoice_line)
             inv.issqn_value = sum(
                 line.issqn_value for line in inv.invoice_line)
@@ -216,7 +200,7 @@ class AccountInvoice(models.Model):
             copy = (0, 0, move[2].copy())
             copy[2]['debit'] = move[2]['credit']
             copy[2]['credit'] = move[2]['debit']
-            copy[2]['name'] = copy[2]['name'] + u'- Retenção'
+            copy[2]['name'] += u'- Retenção'
             return copy
 
         for move in move_lines_tax:
