+++ l10n_br_account_withholding/models/res_company.py
@@ -1,23 +1,7 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    Copyright (C) 2014 KMEE (http://www.kmee.com.br)
-#    @author Luis Felipe Mileo <mileo@kmee.com.br>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# -*- coding: utf-8 -*-
+# © 2016 KMEE(http://www.kmee.com.br)
+#   @author Luis Felipe Mileo <mileo@kmee.com.br>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from openerp import fields, models
 from openerp.addons import decimal_precision as dp
@@ -49,7 +33,9 @@ class ResCompany(models.Model):
         u'Valor mínimo INSS', default=10.0,
         digits_compute=dp.get_precision('Account'))
     irrf_wh_percent = fields.Float(
-        u'Taxa de IR(%)', digits_compute=dp.get_precision('Discount'), default=11.0)
+        u'Taxa de IR(%)',
+        digits_compute=dp.get_precision('Discount'),
+        default=11.0)
     irrf_wh = fields.Boolean(u'Retém IRRF')
     issqn_wh = fields.Boolean(u'Retém ISSQN')
     inss_wh = fields.Boolean(u'Retém INSS')
